+++ sdk/identity/identity/src/client/identityClient.ts
@@ -12,7 +12,7 @@ import {
   createPipelineFromOptions,
   isNode
 } from "@azure/core-http";
-import { CanonicalCode } from "@opentelemetry/types";
+import { CanonicalCode } from "@opentelemetry/api";
 import { AuthenticationError, AuthenticationErrorName } from "./errors";
 import { createSpan } from "../util/tracing";
 import { logger } from "../util/logging";