+++ sdk/identity/identity/src/util/tracing.ts
@@ -3,7 +3,7 @@
 
 import { GetTokenOptions } from "@azure/core-http";
 import { getTracer, OperationTracingOptions } from "@azure/core-tracing";
-import { Span, SpanKind, SpanOptions } from "@opentelemetry/types";
+import { Span, SpanKind, SpanOptions as OTSpanOptions } from "@opentelemetry/api";
 
 /**
  * Creates a span using the global tracer.
@@ -21,7 +21,7 @@ export function createSpan(
     ...options.tracingOptions
   };
 
-  const spanOptions: SpanOptions = {
+  const spanOptions: OTSpanOptions = {
     ...tracingOptions.spanOptions,
     kind: SpanKind.INTERNAL
   };
@@ -37,7 +37,7 @@ export function createSpan(
         ...tracingOptions,
         spanOptions: {
           ...tracingOptions.spanOptions,
-          parent: span,
+          parent: span.context(),
           attributes: {
             ...spanOptions.attributes,
             "az.namespace": "Microsoft.AAD"