+++ sdk/identity/identity/test/node/authorizationCodeCredential.spec.ts
@@ -100,7 +100,7 @@ describe("AuthorizationCodeCredential", function() {
     await credential.getToken("scope", {
       tracingOptions: {
         spanOptions: {
-          parent: rootSpan
+          parent: rootSpan.context()
         }
       }
     });