+++ sdk/storage/storage-file-share/test/fileclient.spec.ts
@@ -682,7 +682,7 @@ describe("FileClient", () => {
     const rootSpan = tracer.startSpan("root");
     await fileClient.create(content.length, {
       tracingOptions: {
-        spanOptions: { parent: rootSpan }
+        spanOptions: { parent: rootSpan.context() }
       }
     });
     rootSpan.end();