+++ sdk/test-utils/recorder/rollup.base.config.js
@@ -84,7 +84,7 @@ export function browserConfig(test = false) {
         // https://github.com/rollup/rollup-plugin-commonjs/issues/394
         namedExports: {
           "events/": ["EventEmitter"],
-          "@opentelemetry/types": ["CanonicalCode", "SpanKind", "TraceFlags"]
+          "@opentelemetry/api": ["CanonicalCode", "SpanKind", "TraceFlags"]
         }
       }),
       viz({ filename: "dist-browser/browser-stats.html", sourcemap: false })