+++ sdk/core/core-http/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Release History
 
-## 1.1.1 (Unreleased)
+## 1.1.1 (2020-04-28)
+
+- Add support for `text/plain` endpoints. [PR #7963](https://github.com/Azure/azure-sdk-for-js/pull/7963)
+- Updated to use OpenTelemetry 0.6.1 via `@azure/core-tracing`.
 
 
 ## 1.1.0 (2020-03-31)